+++ spec/spec_helper.rb
@@ -26,7 +26,7 @@ require_relative 'support/helpers/requires'
 Dir.glob( "#{support_path}/{lib,helpers,shared}/**/*.rb" ).each { |f| require f }
 
 # Uncomment to show output from spawned processes.
-#Arachni::Processes::Manager.preserve_output
+Arachni::Processes::Manager.preserve_output
 
 RSpec.configure do |config|
     config.treat_symbols_as_metadata_keys_with_true_values = true